requirements.txt
pytest
requests
jsonschema
allure-pytest==2.13.2
curlify==2.2.1